Move ModelUrl to its own file
